Minor tweaks

+++ src/p_inter.c
@@ -2388,7 +2388,7 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
 
         if (joy_rumble_damage)
         {
-            const int   strength = (30000 + (100 - MIN(tplayer->health, 100)) / 100 * 30000) * joy_rumble_damage / 100;
+            const short strength = (30000 + (100 - MIN(tplayer->health, 100)) / 100 * 30000) * joy_rumble_damage / 100;
 
             I_GameControllerRumble(strength, strength);
             damagerumbletics += BETWEEN(12, damage, 100);
